requirements.txt
# Core API
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2

# NLP
spacy==3.7.4

# Templating
jinja2==3.1.3

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Airflow
apache-airflow==2.9.1

# Utilities
python-dotenv==1.0.1
